+++ resources/views/wishlist/show.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('My Wishlist')">
+    <livewire:storefront.wishlist-page />
+</x-layouts::storefront>